requirements.txt
streamlit
opencv-python-headless
pandas
Pillow